Bare minimum.
The barest of minimums of what might be considered an implementation of
PGL(2,Z) as a Coxeter group have been done now.
You can do

    sage: W = PGL2Z()
    sage: H = IwahoriHeckeAlgebra(ZZ, 0, 0, W)
    sage: T = H.T()

but unforunately

    sage: T.algebra_generators()

will fail. The reason for this appears to lie in the way the algebra `T` is
constructed as a `FreeModule`. More specifically, the method
`CombinatorialFreeModule.monomial()` seems to be to blame.